separate weatherbit api from general.py
