+++ scripts/general.py
@@ -1,16 +1,13 @@
 # Source: https://www.kaggle.com/datasets/khsamaha/aviation-accident-database-synopses
 
-from datetime import timedelta
 
-import config
 import matplotlib.pyplot as plt
-import numpy as np
 import pandas as pd
-import requests
 import seaborn as sns
 from load_airplane_accidents_dataset import load_dataset
 from preprocesse_dataset import preprocese_dataset
 from utils import logger_df
+from weatherbit_api import add_data_from_weatherbit_api
 
 
 def filter_by_period(
@@ -132,41 +129,6 @@ def plot_accidents_per_year(df_accidents_per_year: pd.DataFrame) -> None:
     sns.lineplot(data=df_accidents_per_year, x="Event_year", y="Count", color="#2990EA")
 
 
-@logger_df
-def add_data_from_weatherbit_api(df: pd.DataFrame) -> pd.DataFrame:
-    """Reads data (temperature by day) from Weatherbit.io regarding accident data"""
-    temperatures = []
-    df = df.tail(5)
-    for row in df.itertuples():
-        params = build_weatherbit_api_params(row)
-        response = api_request_weatherbit_api(params)
-        temperatures.append(response)
-    return df.assign(Temperatures_accident_day=temperatures)
-
-
-def build_weatherbit_api_params(row) -> dict:
-    start_date = row.Event_Date.date()
-    end_date = start_date + timedelta(days=1)
-    params = {
-        "key": [config.WEATHERBIT_API_KEY],
-        "start_date": start_date,
-        "end_date": end_date,
-        "city": row.City,
-    }
-    return params
-
-
-def api_request_weatherbit_api(parameters):
-    method = "ping"
-    api_base = "https://api.weatherbit.io/v2.0/history/daily?"
-    try:
-        api_result = requests.get(api_base + method, parameters)
-        response = api_result.json()
-    except ValueError:
-        return np.nan
-    return response["data"][0]["temp"]
-
-
 if __name__ == "__main__":
     df = load_dataset()
     df_processed = preprocese_dataset(df)
@@ -182,7 +144,4 @@ if __name__ == "__main__":
 
     accident_statistics_by_airplane_make_engine_flight_purpose(df_processed)
 
-    # df_with_external_data = add_data_from_weatherbit_api(df_processed)
-
-    plot_accidents_per_year(incidents_per_year)
-    plt.show()
+    df_with_external_data = add_data_from_weatherbit_api(df_processed)
